Worked on Web Structure and added Library
The goal of the push was to try and improve website structure.

Changed the name of the "Tutrorial" page to the Library Page because I thought it better reflected upon the promise I wish to Deliver visitors.

Added a Web Technologies Category to the Library. This will reflect the the knowledge and experience I accrue as I work with HTML, CSS, Javascript, SQL and Azure.