Restructure and Clean up
Restructured the file Hierarchy for a more organized pattern for finding libraries and keeping files organized in a logical manner.

I Cleaned up a the startup Program.cs file so it's not so sloppy. I added error checking for retrieving azure secrets.

I added a database and verified I was able to connect to it through my webapp.