AppSetting.json Update
I was able follow microsoft documentation to enable logging on azure which revealed that the program.cs was throwing an exception at line 14. Which was the list of strings I created out of appsettings file.

The issue in the appsetting.json file.

I had been working in the appsettings.development.json file and didn't apply the changes to the appsettings.json file.

I corrected the file.