Basic Working Prototype for settings form
Updated Database Tables with better user identifiers, and also added a table to show user privileges.
Privileges will be used to vet posts made, or give access to manipulate some data in the database.

This update is just a prototype where I Updated a record in the database for the first time through a razorpage.

The code committed is heading towards a refactor of the UserService claass.

Right now I'm passing the user ID as a parameter in a function. I don't like this because I'm concerned that value can be changed.  I'm doing some research to figure out how I can implement a different solution and add some checks to ensure the person updating the database is the person the authentication says they should be.