Add Moderator Function to Update Library Quick Links
I Created the functionality to allow a Moderator to update the quick links in the library.

These quick links are stored in the database as a single Json string.

The User interface caps the number of links that can be stored to 5 quick links.

I Created a LibraryLinksController to communicate with the LibraryService class.
I reuse the same functions to populate the form to edit the link data, except I use a switch to select the view I want to pass the data in.