Finalized Video Post Prototype
Added  functionality, query and functions need to to retrieve a list of links from the database, which are then used to create video previews.

Updated the Color System javascript to populate the pickers with dark mode color scheme.

Commented out Language Support code because it can't do anything right now but break things.