Refactoring Nearing Completion.
Renamed the QueryBuilder to QueryHelper because it QueryBuiilder was already a class used by the .net framwork.

Project is Organized with more intuitive naming conventions.

The base Razorpage class forces the use of the topic id in the constructor. This will simplify find data in the database and I used this value to write simpler queries.

Temporarily took out logging in as a requirement to test the features.