Implemented Picture in Picture
Created a area to host an instance of my website in an iframe inside my website.

To get an instance of my website to work, it was mostly finding and copying the elements the the correct ID's on it into the new Layout_iFrame Layout I made to be used in the iframe. The id's in the iframe remain completely isolated from the rest of the site.