Integrated Language Support
Created language tables to house alias names for the database map structure and figured out how to communicate the user's chosen language in the BaseRazorPage Model.

I added some simple controls to that allows the user alter the alias names of the folders and razor pages.

Created data driven language option from the menu, inspector windows and the mini view and allowed the map selectors to populate with the alias names. The menu Language and mini view languages have been decoupled and can now load different languages.

What i need to do next is when I add a new language or map structure to the database, it needs to run some code to initialize the respective areas that enables support for those languages.

The basic system to include language support should allow data driven explanations to be included into the razor pages.